Give the placeholder-order pins their own file (refs #439)
Follow-up to the review fixes. The new pin pushed `HTMLEscapingTests` past
its type-body limit, and the failure it pins is not an escaping failure at
all: every value in the chain is correctly escaped and still changes the
markup around it, because a chain of replacements fills the placeholders an
earlier replacement inserted. `PlaceholderOrderTests` says that in its name.

Adds the boundary the reorder does not move. Two of the picker's
placeholders carry test-plan text and no order protects both; the one left
reachable is `[[CURRENT_DEVICE]]`, deliberately, because what it inserts is a
destination label rather than the panel of buttons the opposite order handed
out. The test asserts exactly that much — a name can garble text, never put
a control in the picker — and names what closing the class outright would
take: single-pass substitution across the whole `HTML` seam, which every
template shares and which predates this work.

Refs #439.
